# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog OpenApi Documentation",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/update-blogpost/{id}": {
            "put": {
                "tags": [
                    "Blog Post"
                ],
                "summary": "update",
                "operationId": "update ",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
        "/api/delete-blogpost/{id}": {
            "delete": {
                "tags": [
                    "Blog Post"
                ],
                "summary": "Delete",
                "operationId": "Delete ",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
...
        "/api/store-product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "store product",
                "operationId": "store product",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "name",
                                    "description",
                                    "price",
                                    "category"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Validations: filetype=jpg,jpeg,png",
                                        "type": "file"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
        "/api/update-product/{id}": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Update product",
                "operationId": "Update product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "name",
                                    "description",
                                    "price",
                                    "category"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Validations: filetype=jpg,jpeg,png",
                                        "type": "file"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
        "/api/delete-product/{id}": {
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product",
                "operationId": "Delete Product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        },
        "/api/view-product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "View Product",
                "operationId": "View Product",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter_data",
                        "in": "query",
                        "description": " Filter data",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "contains",
                            "enum": [
                                "contains",
                                "notContains",
                                "equals",
                                "notEqual",
                                "startsWith",
                                "endsWith",
                                "inRange"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "APIKEY": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "APIKEY": {
                "type": "apiKey",
                "description": "Default Token :static",
                "name": "APIKEY",
                "in": "header"
            },
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}